Allow coupons to be combined
Previously the date format was set to `Ymds` which allowed generated coupons with a unique hash to be created multiple times. A user could click the apply credit button and generate multiple coupons, applying them and totaling a greater value than the store credit they had to use. 

This commit changes the date format to `YmdHi` switching seconds for hour and minutes. Now if you click the apply coupon multiple times, the first coupon will generate and any subsequent coupons will fail to generate since a coupon with the same unique id already exists. 

An edge case still exists where the person could do this right when the minute changes. At most it would allow them 2 generated coupons while the server was responding. This would require the customer to know the exact server time to start clicking and span the 59-00 second/minute change. In testing I was unable to successfully accomplish this. In looking at this more objectively, this could also apply if you reduced to hours, or days. There always exists a couple second point of time where the server is loading and a minute to minute, hour to hour, or day to day change is happening. It's a race condition that can't be beat, but has a low probability of being exploited successfully. 

Fixes #67 #74